Add files via upload
On October 31st, 2023, I took the Integrity Pledge as part of the fight against corruption, an initiative under the Central Vigilance Commission.
I firmly believe that integrity is the foundation of a just and progressive society. By pledging to follow probity, uphold transparency, and act in the public interest, I reaffirm my responsibility not just as a student — but as a citizen. 🇮🇳
Let’s lead by example, report wrongs, and inspire change — because ethical values shape powerful futures.