Add files via upload
"Innovation begins with curiosity and the courage to showcase it!"
Proud to have represented the ECE Department of Sri Manakula Vinayagar Engineering College at Innovators Day 2024, a National Level Project Contest held on 13.09.2024.
This platform provided a great opportunity to display creativity, technical skills, and teamwork in front of academic and industry experts. A valuable step forward in my engineering journey!